Quick test refactor
