+++ test/__init__.py
@@ -6,6 +6,8 @@ import bson
 import bsonnumpy
 import numpy as np
 import pymongo
+from bson.codec_options import CodecOptions
+from bson.raw_bson import RawBSONDocument
 
 if sys.version_info[:2] == (2, 6):
     import unittest2 as unittest
@@ -69,8 +71,12 @@ class ClientContext(object):
 client_context = ClientContext()
 
 
-class TestFromBSON(unittest.TestCase):
-    def compare_results(self, np_type, document, compare_to):
+class TestToNdarray(unittest.TestCase):
+    @classmethod
+    def setUpClass(cls):
+        cls.client = client_context.client
+
+    def compare_array_results(self, np_type, document, compare_to):
         data = bson._dict_to_bson(document, False, bson.DEFAULT_CODEC_OPTIONS)
         dtype = np.dtype(np_type)
         result = bsonnumpy.bson_to_ndarray(data, dtype)
@@ -80,6 +86,56 @@ class TestFromBSON(unittest.TestCase):
                              "Comparison failed for type %s: %s != %s" % (
                                  dtype, compare_to[str(i)], result[i]))
 
+    def compare_elements(self, expected, actual, dtype):
+        if isinstance(expected, dict):
+            for key, value in expected.items():
+                self.compare_elements(value, actual[key],
+                                      dtype=dtype.fields[key][0])
+
+        elif isinstance(expected, list):
+            self.assertEqual(len(actual), len(expected))
+
+            # If an array's shape is (3,2), its subarrays' shapes are (2,).
+            subdtype, shape = dtype.subdtype
+            self.assertGreaterEqual(len(shape), 1)
+            subarray_dtype = np.dtype((subdtype, shape[1:]))
+            for i in range(len(expected)):
+                self.compare_elements(expected[i], actual[i], subarray_dtype)
+
+        elif dtype.kind == 'V':
+            self.assertEqual(bytes(expected.ljust(dtype.itemsize, b'\0')),
+                             bytes(actual))
+
+        elif dtype.kind == 'S':
+            # NumPy only stores bytes, not str.
+            self.assertEqual(expected, actual.decode('utf-8'))
+
+        else:
+            self.assertEqual(expected, actual)
+
+    # TODO: deal with both name and title in dtype
+    def compare_results(self, dtype, expected, actual):
+        self.assertEqual(dtype, actual.dtype)
+        self.assertEqual(expected.count(), len(actual))
+        for act in actual:
+            exp = next(expected)
+            for name in dtype.names:
+                self.compare_elements(exp[name], act[name],
+                                      dtype=dtype.fields[name][0])
+
+    def make_mixed_collection_test(self, docs, dtype):
+        self.client.bsonnumpy_test.coll.delete_many({})
+        self.client.bsonnumpy_test.coll.insert_many(docs)
+        raw_coll = self.client.get_database(
+            'bsonnumpy_test',
+            codec_options=CodecOptions(document_class=RawBSONDocument)).coll
+        cursor = raw_coll.find()
+
+        ndarray = bsonnumpy.sequence_to_ndarray(
+            (doc.raw for doc in cursor), dtype, raw_coll.count())
+        self.compare_results(np.dtype(dtype),
+                             self.client.bsonnumpy_test.coll.find(),
+                             ndarray)
 
 def millis(delta):
     if hasattr(delta, 'total_seconds'):
